+++ src/index.js
@@ -6,6 +6,13 @@ export function isSupported() {
   return typeof Proxy !== 'undefined';
 }
 
+function shouldValueWrappedAsProxy(value) {
+  if (value == null) return false;
+  if (typeof value !== 'object') return false;
+  if (value instanceof Date) return false;
+  return true;
+}
+
 const defaultCompare = (a, b) => a === b;
 
 export default function reactiveProxy(initial = {}, compare, changeCallback) {
@@ -38,7 +45,7 @@ export default function reactiveProxy(initial = {}, compare, changeCallback) {
         }
         if (changeCallback) changeCallback(key, value);
       };
-      obj[key] = options.recursive && typeof value === 'object' && value != null
+      obj[key] = options.recursive && shouldValueWrappedAsProxy(value)
         ? reactiveProxy(value, options, changed)
         : value;